Move prognostic solver tests
